refactor: remove deprecated singer tap
