+++ backend/core/models/migrationscripts/20230822_drop_tap_state.go
@@ -15,11 +15,26 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package models
+package migrationscripts
 
-// The consts that this plugin needs
-const (
-	TapExecutable        = "tap-pagerduty"
-	StreamPropertiesFile = "pagerduty.json"
-	IncidentStream       = "incidents"
+import (
+	"github.com/apache/incubator-devlake/core/context"
+	"github.com/apache/incubator-devlake/core/errors"
+	"github.com/apache/incubator-devlake/core/plugin"
 )
+
+var _ plugin.MigrationScript = (*dropTapStateTable)(nil)
+
+type dropTapStateTable struct{}
+
+func (script *dropTapStateTable) Up(basicRes context.BasicRes) errors.Error {
+	return basicRes.GetDal().DropTables("_devlake_collector_tap_state")
+}
+
+func (*dropTapStateTable) Version() uint64 {
+	return 20230822145125
+}
+
+func (*dropTapStateTable) Name() string {
+	return "drop _devlake_collector_tap_state table"
+}
